Added support for static library builds. Whether the build is static or not is controlled with BOOST_LOG_DYN_LINK and BOOST_ALL_DYN_LINK macros. The former, if defined, makes Boost.Log to be dynamically linked, the latter makes the same for all Boost libraries. If neither of them is defined, the library is linked statically.
Added support for configuration macros BOOST_LOG_USE_CHAR and BOOST_LOG_USE_WCHAR_T. If either of them is defined, the corresponding character support is enabled. If neither is defined (which is the default), it is assumed that both are in effect. This allows to reduce binary size to half of the full size.


git-svn-id: https://boost-log.svn.sourceforge.net/svnroot/boost-log/trunk/boost-log@109 ab6a8de6-4e00-4c3d-a860-8899c17d353f